Move scheduler to core module
